requirements.txt
Flask==3.1.0
python-docx==1.1.2
openpyxl==3.1.5
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.0
pdfplumber==0.11.4
reportlab==4.2.5
cryptography==42.0.0
